Refactor navbar with search
